tpm: Move TPM2 specific definitions and functions to new headers
This gathers all the TPM2 definitions and structures into two separate
header files (public tpm2.h and private tpm2_structs.h). The definitions
moved to these files correspond to the TCG specification for TPM 1 family:

TPM 2.0 Library
 - https://trustedcomputinggroup.org/resource/tpm-library-specification/

Note that the structures were pulled into tpm2_structs.h to allow their
external reuse.
